NVIM: move configs in a namespace to avoid potential conflicts with lua runtime
