+++ install.conf.yaml
@@ -8,6 +8,7 @@
   - ~/.config/powerline-shell
   - ~/.config/nvim
   - ~/.config/nvim/lua
+  - ~/.config/nvim/lua/iamgianluca
   - ~/.config/nvim/after/plugin
   - ~/.dwm
   - ~/.zfunc
@@ -18,11 +19,11 @@
     ~/.zfunc/_poetry: poetry/_poetry
     ~/.config/powerline-shell/config.json: powerline_shell/powerline-shell.json
     ~/.config/nvim/init.lua: nvim/init.lua
-    ~/.config/nvim/lua/remap.lua: nvim/lua/remap.lua
-    ~/.config/nvim/lua/set.lua: nvim/lua/set.lua
-    ~/.config/nvim/lua/plugins.lua: nvim/lua/plugins.lua
+    ~/.config/nvim/lua/iamgianluca/remap.lua: nvim/lua/iamgianluca/remap.lua
+    ~/.config/nvim/lua/iamgianluca/set.lua: nvim/lua/iamgianluca/set.lua
+    ~/.config/nvim/lua/iamgianluca/plugins.lua: nvim/lua/iamgianluca/plugins.lua
+    ~/.config/nvim/lua/iamgianluca/theme.lua: nvim/lua/iamgianluca/theme.lua
     ~/.config/nvim/after/plugin/treesitter.lua: nvim/after/plugin/treesitter.lua
-    ~/.config/nvim/lua/theme.lua: nvim/lua/theme.lua
     ~/.config/nvim/after/plugin/telescope.lua: nvim/after/plugin/telescope.lua
     ~/.config/nvim/after/plugin/lsp.lua: nvim/after/plugin/lsp.lua
     ~/.config/nvim/after/plugin/comment.lua: nvim/after/plugin/comment.lua
